feat: 移除 v2 路由與跳轉連結
