+++ pages/agenda/index.vue
@@ -1,285 +1,94 @@
 <script setup lang="ts">
-const router = useRouter()
-const route = useRoute()
+import type { AgendaItem, AgendaTag } from '~/types'
+import { AGENDA_LIST } from '~/constants/agenda'
 
 useSeoMeta({
   title: '議程資訊',
 })
 
-type AgendaTag
-  = | 'AI'
-    | 'Frontend'
-    | 'Backend'
-    | 'Security'
-    | '軟體設計'
-    | '設計實務'
-    | '產品思維'
-    | '產業應用'
-    | '團隊管理'
-    | 'Agile'
-
-interface SpeakerInfo {
-  name: string
-  title: string
+// 將議程按時間分組
+function groupAgendasByTime(agendas: AgendaItem[]) {
+  return agendas.reduce(
+    (acc, agenda) => {
+      const time = agenda.startTime
+      if (!acc[time]) {
+        acc[time] = []
+      }
+      acc[time].push(agenda)
+      return acc
+    },
+    {} as Record<string, AgendaItem[]>,
+  )
 }
 
-interface AgendaItem {
-  topic: string
-  speakerInfo: SpeakerInfo[]
-  tags: AgendaTag[]
-}
+const agendasAtDec12Morning = computed(() => {
+  const morningAgendas = AGENDA_LIST.filter(
+    item => item.day === '12' && item.startTime < '12:00',
+  )
+  return groupAgendasByTime(morningAgendas)
+})
 
-const AGENDA_LIST: AgendaItem[] = [
-  {
-    topic: 'B2B 服務的 AI Agent 產品設計原則',
-    speakerInfo: [
-      {
-        name: 'Happy',
-        title: '91APP / 產品長',
-      },
-    ],
-    tags: ['AI', '產品思維', '產業應用', '團隊管理'],
-  },
-  {
-    topic: '從冷知識到漏洞：你不懂的 Web，駭客懂',
-    speakerInfo: [
-      {
-        name: '胡立',
-        title: `技術部落格 Huli's blog / 站長`,
-      },
-    ],
-    tags: ['Frontend', 'Backend', 'Security'],
-  },
-  {
-    topic: '程式碼與尿布：媽媽工程師的生存指南',
-    speakerInfo: [
-      {
-        name: 'Hannah',
-        title: 'Remote / Frontend ',
-      },
-    ],
-    tags: ['Frontend'],
-  },
-  {
-    topic: '大 AI 時代，工程師的成長之路 / 從 Junior 到 Staff',
-    speakerInfo: [
-      {
-        name: '奶綠茶',
-        title: 'PositiveGrid Staff Frontend Engineer',
-      },
-    ],
-    tags: ['Frontend'],
-  },
-  {
-    topic: '別再瞎忙了！讓 AI 幫產品團隊找到對的問題',
-    speakerInfo: [
-      {
-        name: 'Peter',
-        title: '漸強實驗室 / Product Lead',
-      },
-    ],
-    tags: ['AI', '產品思維', '團隊管理', 'Agile'],
-  },
-  {
-    topic: 'AI 時代下， Product Sense 就是你的秘密武器',
-    speakerInfo: [
-      {
-        name: 'Chris',
-        title: 'PicCollage 拼貼趣 <br /> / 資深產品設計經理',
-      },
-      {
-        name: 'Tiyna',
-        title: 'PicCollage 拼貼趣 <br /> / 資深產品設計師',
-      },
-    ],
-    tags: ['AI', '設計實務', '產品思維'],
-  },
-  {
-    topic: '軟體開發邪教的救贖：AI 時代更應掌握的 TDD 技能',
-    speakerInfo: [
-      {
-        name: 'Kuma Syu',
-        title: '緯雲有限公司 / R&D Manager',
-      },
-    ],
-    tags: ['AI', '軟體設計', '產品思維'],
-  },
-  {
-    topic: '零基礎打造 400 萬用戶，我們犯了哪些錯誤',
-    speakerInfo: [
-      {
-        name: 'Max Chen',
-        title: '對稱資訊股份有限公司 / 面試趣 CEO',
-      },
-    ],
-    tags: ['產品思維', '產業應用'],
-  },
-  {
-    topic: '產品 OKR 的訂立與 Roadmap 展開',
-    speakerInfo: [
-      {
-        name: '曾友志',
-        title: '下午先生有限公司 / 資深產品顧問',
-      },
-    ],
-    tags: ['產品思維', '團隊管理'],
-  },
-  {
-    topic: 'React 優化實戰分析 / 掌握 React 進階技術 x 底層思維',
-    speakerInfo: [
-      {
-        name: 'ThisWeb (Kun)',
-        title: 'ThisWeb / 前端工程師',
-      },
-    ],
-    tags: ['Frontend'],
-  },
-  {
-    topic: '願 Web API 原力與你同在',
-    speakerInfo: [
-      {
-        name: 'MUKI',
-        title: 'MUKI space* / 前端工程師',
-      },
-    ],
-    tags: ['AI', 'Frontend'],
-  },
-  {
-    topic: '掌握田野中的「人」：真實場域研究的人際溝通與信任建立',
-    speakerInfo: [
-      {
-        name: 'Joey',
-        title: '引鹿創新體驗研究室  / UX 總監',
-      },
-    ],
-    tags: ['設計實務', '團隊管理'],
-  },
-  {
-    topic: '以打詐為例，服務設計如何讓公共數位服務有感',
-    speakerInfo: [
-      {
-        name: '卓致遠',
-        title: '致遠體驗設計 / 體驗總監',
-      },
-    ],
-    tags: ['設計實務', '產品思維'],
-  },
-  {
-    topic: '服務設計師的傳說：90% 的時間在讓設計發生',
-    speakerInfo: [
-      {
-        name: '郭心喻',
-        title:
-          'Government of Nova Scotia 加拿大新斯科舍省政府 <br /> / Senior Service Designer 資深服務設計師',
-      },
-    ],
-    tags: ['軟體設計', '設計實務', '產品思維', '產業應用'],
-  },
-  {
-    topic: '從設計到共識：悠識如何在每一次新專案裡，用溝通建立信任關係',
-    speakerInfo: [
-      {
-        name: '林星妤',
-        title: '悠識數位 / Senior Experience Designer',
-      },
-      {
-        name: '黃明硯',
-        title: '悠識數位 /  Experience Designer',
-      },
-    ],
-    tags: ['設計實務', '產品思維'],
-  },
-  {
-    topic: '從研究到上線：全流程設計與 AI 協作專案實戰',
-    speakerInfo: [
-      {
-        name: 'Niki Liu',
-        title: '新加坡商鈦坦科技 / Senior Product Designer',
-      },
-    ],
-    tags: ['AI', '設計實務', '產品思維'],
-  },
-  {
-    topic: '從使用者到工程師：AI 在企業怎麼落地？',
-    speakerInfo: [
-      {
-        name: 'Shirney Huang 黃琇琳',
-        title: 'Aiworks / CEO',
-      },
-    ],
-    tags: ['AI'],
-  },
-  {
-    topic: 'GenAI 時代下的測試三板斧',
-    speakerInfo: [
-      {
-        name: '柯仁傑',
-        title: 'Odd-e / Technical Coach',
-      },
-    ],
-    tags: ['軟體設計'],
-  },
-  {
-    topic: '培養十倍速 PM：從工具到習慣',
-    speakerInfo: [
-      {
-        name: '趙柏強',
-        title: 'Stable Progress / Growth Companion',
-      },
-    ],
-    tags: ['AI', '產品思維', '團隊管理', 'Agile'],
-  },
-  {
-    topic: '對微前端的美好想像',
-    speakerInfo: [
-      {
-        name: 'Eric Lee',
-        title: '前端輕鬆聊 / Senior Software Engineer',
-      },
-    ],
-    tags: ['Frontend', '軟體設計'],
-  },
-  {
-    topic: 'AI 時代下， 如何運用數據幫助產品發揮影響力',
-    speakerInfo: [
-      {
-        name: '姜乃文',
-        title: '天下雜誌集團 / 主任分析師',
-      },
-    ],
-    tags: ['AI', '產品思維', '產業應用', '團隊管理'],
-  },
-  {
-    topic: '走在同業前面：國泰的雲端轉型洞察與啟示',
-    speakerInfo: [
-      {
-        name: '顏勝豪 Otto',
-        title: '國泰金控  / 協理',
-      },
-    ],
-    tags: ['AI', '產業應用'],
-  },
-]
+const agendasAtDec12Afternoon = computed(() => {
+  const afternoonAgendas = AGENDA_LIST.filter(
+    item => item.day === '12' && item.startTime >= '12:00',
+  )
+  return groupAgendasByTime(afternoonAgendas)
+})
 
-const isShowPopover = ref(false)
+const agendasAtDec13Morning = computed(() => {
+  const morningAgendas = AGENDA_LIST.filter(
+    item => item.day === '13' && item.startTime < '12:00',
+  )
+  return groupAgendasByTime(morningAgendas)
+})
 
-function handlePopoverShow(topic: string) {
-  isShowPopover.value = true
+const agendasAtDec13Afternoon = computed(() => {
+  const afternoonAgendas = AGENDA_LIST.filter(
+    item => item.day === '13' && item.startTime >= '12:00',
+  )
+  return groupAgendasByTime(afternoonAgendas)
+})
 
-  router.push({
-    query: { ...route.query, speakerId: topic },
-  })
-}
+const selectedDate = ref(new Date() <= new Date('2025-12-13') ? '12' : '13')
 
-watch(
-  () => route.query.speakerId,
-  (newVal) => {
-    if (newVal && typeof newVal === 'string') {
-      handlePopoverShow(newVal)
-    }
-  },
-  { immediate: true },
+// 監聽 selectedDate 變化，切換時滾動到頂部
+watch(selectedDate, () => {
+  const lenis = useLenis()
+  if (lenis) {
+    lenis.scrollTo(0)
+  }
+})
+
+// 根據選定日期動態獲取議程
+const currentMorningAgendas = computed(() =>
+  selectedDate.value === '12'
+    ? agendasAtDec12Morning.value
+    : agendasAtDec13Morning.value,
 )
+
+const currentAfternoonAgendas = computed(() =>
+  selectedDate.value === '12'
+    ? agendasAtDec12Afternoon.value
+    : agendasAtDec13Afternoon.value,
+)
+
+const isMenuOpen = ref(false)
+const showAside = ref(false)
+
+const selectedTags = ref<AgendaTag[]>([])
+
+// 延遲隱藏 aside 以確保過渡完成
+watch(isMenuOpen, (newValue) => {
+  if (newValue) {
+    showAside.value = true
+  }
+  else {
+    // 等待 leave transition 完成後再隱藏
+    setTimeout(() => {
+      showAside.value = false
+    }, 200)
+  }
+})
 </script>
 
 <template>
@@ -288,6 +97,7 @@ watch(
       <div class="agenda-section relative h-80 text-webconf-gray lg:h-400">
         <!-- 浮動方框 -->
         <ShareLayoutBlocks />
+
         <!-- 桌機使用 -->
         <svg
           class="absolute inset-0 hidden size-full sm:block"
@@ -312,7 +122,7 @@ watch(
             <!-- 靜態軌道 -->
             <line
               x1="0"
-              y1="55"
+              y1="54.5"
               x2="345"
               y2="400"
               stroke="#E6E6E6"
@@ -345,105 +155,265 @@ watch(
           class="absolute left-[50%] top-[45%] flex w-full -translate-x-1/2 flex-col items-center justify-center gap-6 px-5 sm:top-[38%] sm:px-6 lg:top-[50%] lg:flex-row lg:items-end"
         >
           <h1 class="text-h1-96 text-white">
-            AGENDA
+            Agenda
           </h1>
           <div class="flex w-full flex-col gap-3 lg:w-fit">
             <span
               class="inline-block px-0 text-center text-h4-24 lg:pl-[285px] lg:pr-10"
             >議程資訊</span>
             <div class="order-[-1] flex items-center lg:order-1">
-              <span class="size-3 bg-white"></span>
-              <span class="h-[1px] flex-1 bg-white"></span>
-              <span class="size-3 bg-white"></span>
+              <span class="size-3 bg-webconf-gray"></span>
+              <span class="h-[1px] flex-1 bg-webconf-gray"></span>
+              <span class="size-3 bg-webconf-gray"></span>
             </div>
           </div>
         </div>
+
         <div
           v-arrow="{ speed1: '12s', color: '#E6E6E6' }"
-          class="absolute bottom-0 left-0 z-20 h-[1px] w-full bg-webconf-gray"
+          class="absolute bottom-0 left-0 z-30 h-[1px] w-full bg-webconf-gray"
         ></div>
       </div>
     </section>
 
-    <main class="text-webconf-gray">
-      <!-- 活動日期 -->
+    <main class="border-b border-webconf-gray text-webconf-gray">
       <section
-        class="sticky top-[55px] z-10 border-b border-webconf-gray bg-black"
+        class="sticky top-[54px] z-20 border-b border-webconf-gray bg-black sm:top-[57px] lg:top-[54px]"
       >
-        <div
-          class="container px-5 py-4 text-h4-24 sm:px-8 sm:py-5 lg:px-20 lg:py-6"
-        >
-          Dec. 12-13
+        <div class="flex-center py-[6px] text-h4-24 lg:py-3">
+          <!-- 日期篩選 -->
+          <AgendaEventDaySwitch v-model="selectedDate" />
+
+          <!-- 議程類型篩選 -->
+          <AgendaTagFilterBtn
+            :is-menu-open="isMenuOpen"
+            size="sm"
+            :selected-tags-count="selectedTags.length"
+            class="ml-3 lg:hidden"
+            @click="isMenuOpen = true"
+          />
+
+          <!-- 廳號 -->
+          <div
+            class="hidden grow text-center text-h4-24 lg:grid lg:grid-cols-3"
+          >
+            <h3 class="col-span-1">
+              A2 棟
+            </h3>
+            <h3 class="col-span-1">
+              M 棟
+            </h3>
+            <h3 class="col-span-1">
+              F 棟
+            </h3>
+          </div>
         </div>
 
         <div
           v-arrow="{ speed1: '10s', color: '#E6E6E6' }"
-          class="relative z-[100]"
+          class="relative z-30 hidden lg:block"
         ></div>
       </section>
 
       <!-- 議程列表 -->
-      <section class="container sm:px-8 lg:px-20">
-        <ul
-          class="border-x-0 border-x-webconf-gray/50 bg-black sm:border-x-[0.5px]"
+      <section class="relative flex">
+        <!-- 議程類型篩選清單 -->
+        <aside
+          :class="{
+            'z-50 flex': isMenuOpen || showAside,
+            'hidden lg:z-auto lg:flex': !isMenuOpen && !showAside,
+          }"
+          class="fixed top-[54.5px] h-[calc(100dvh-54px)] w-0 shrink-0 flex-col items-start self-start border-webconf-gray bg-black lg:sticky lg:top-[124.5px] lg:w-[228px] lg:border-r lg:p-4 2xl:w-[260px] 2xl:pl-12"
         >
-          <li
-            v-for="item in AGENDA_LIST"
-            :key="item.topic"
-            class="flex flex-col gap-10 border-b-[0.5px] border-b-webconf-gray px-5 pb-10 pt-6 sm:flex-row sm:justify-between sm:px-10 sm:pb-[68px] sm:pt-8"
-            @click="handlePopoverShow(item.topic)"
-          >
-            <!-- 議程主題 -->
-            <div class="sm:flex-[3] lg:flex-[5]">
-              <h2 class="text-h3-40 leading-[1.1]">
-                {{ item.topic }}
-              </h2>
-              <div class="mt-4 flex flex-wrap gap-3 sm:mt-5">
-                <span
-                  v-for="tag in item.tags"
-                  :key="tag + item.topic"
-                  class="gap-3 bg-webconf-blue px-4 py-[6px] text-xs font-semibold text-webconf-gray"
-                >
-                  {{ tag }}
-                </span>
-              </div>
-            </div>
+          <!-- 篩選按鈕 -->
+          <AgendaTagFilterBtn
+            :is-menu-open="isMenuOpen"
+            size="lg"
+            :selected-tags-count="selectedTags.length"
+            @click="isMenuOpen = true"
+          />
 
-            <!-- 講者資訊 -->
-            <ul
-              class="flex w-full gap-6 sm:flex-1 sm:flex-col sm:self-end lg:min-w-[200px]"
+          <!-- 篩選清單 -->
+          <AgendaTagFilterMenu
+            v-model:is-open="isMenuOpen"
+            v-model:selected-tags="selectedTags"
+          />
+
+          <!-- 半透明遮罩 -->
+          <transition
+            enter-active-class="transition-opacity duration-300"
+            leave-active-class="transition-opacity duration-200"
+            enter-from-class="opacity-0"
+            enter-to-class="opacity-100"
+            leave-from-class="opacity-100"
+            leave-to-class="opacity-0"
+          >
+            <div
+              v-if="isMenuOpen"
+              class="absolute inset-0 z-10 h-full w-[100dvw] bg-black/80"
+              @click="isMenuOpen = false"
+            ></div>
+          </transition>
+        </aside>
+
+        <div
+          v-cursor="{
+            scale: 0.5,
+            duration: 0.5,
+          }"
+          class="grid grow grid-cols-3 bg-black"
+        >
+          <transition-group
+            name="agenda-fade"
+            tag="div"
+            class="col-span-3"
+            appear
+          >
+            <div
+              v-if="selectedDate === '12'"
+              class="relative"
             >
-              <li
-                v-for="speaker in item.speakerInfo"
-                :key="speaker.name"
-                class="w-full"
+              <time
+                class="sticky top-[99px] z-[5] flex h-7 w-full items-center bg-webconf-gray px-5 text-btn-16 text-webconf-blue lg:top-[124px] lg:ml-[-86px] lg:w-[86px] lg:justify-center"
               >
-                <h3 class="text-h4-24 leading-[1.1]">
-                  {{ speaker.name }}
-                </h3>
-                <p
-                  class="mt-2 text-xs"
-                  v-html="speaker.title"
-                ></p>
-              </li>
-            </ul>
-          </li>
-        </ul>
+                <span>09:00</span>
+                <span class="lg:hidden">{{ ` - ` }}</span>
+                <span class="lg:hidden">09:10</span>
+              </time>
+
+              <h2
+                key="opening-speech"
+                class="col-span-3 border-b-[0.5px] border-webconf-gray/50 bg-black px-5 py-7 text-h4-24 lg:py-10 lg:text-center xl:py-14"
+              >
+                開幕致詞
+              </h2>
+            </div>
+          </transition-group>
+
+          <transition-group
+            name="agenda-fade"
+            tag="div"
+            class="col-span-3"
+            appear
+          >
+            <div
+              v-for="(agendas, time) in currentMorningAgendas"
+              :key="`morning-${selectedDate}-${time}`"
+              class="relative grid grid-cols-1 lg:grid-cols-3"
+            >
+              <!-- 時間標記 (行動版) -->
+              <div
+                class="sticky top-[99px] z-[5] flex h-7 w-full items-center bg-webconf-gray px-5 text-btn-16 text-webconf-blue xs:top-[107px] sm:top-[109px] lg:top-[107px] lg:hidden"
+              >
+                <time :datetime="time">
+                  {{ time }}
+                </time>
+                <span>{{ ` - ` }}</span>
+                <time :datetime="agendas[0].endTime">
+                  {{ agendas[0].endTime }}
+                </time>
+              </div>
+
+              <AgendaCard
+                v-for="(agenda, index) in agendas"
+                :key="`${time}-${index}`"
+                :data="agenda"
+                :index="index"
+                :time="time"
+                :show-time="index === 0"
+                :is-selected="
+                  agenda.tags?.some((tag) => selectedTags.includes(tag))
+                    || selectedTags.length === 0
+                "
+              />
+            </div>
+          </transition-group>
+
+          <transition-group
+            name="agenda-fade"
+            tag="div"
+            class="col-span-3"
+            appear
+          >
+            <h2
+              key="lunch-break"
+              class="col-span-3 border-b-[0.5px] border-webconf-gray/50 bg-black px-5 py-7 text-h4-24 lg:py-10 lg:text-center xl:py-14"
+            >
+              午休時間
+            </h2>
+          </transition-group>
+
+          <transition-group
+            name="agenda-fade"
+            tag="div"
+            class="col-span-3"
+            appear
+          >
+            <div
+              v-for="(agendas, time) in currentAfternoonAgendas"
+              :key="`afternoon-${selectedDate}-${time}`"
+              class="relative grid grid-cols-1 lg:grid-cols-3"
+            >
+              <!-- 時間標記 (行動版) -->
+              <div
+                class="sticky top-[99px] z-[5] flex h-7 w-full items-center bg-webconf-gray px-5 text-btn-16 text-webconf-blue xs:top-[107px] sm:top-[109px] lg:top-[107px] lg:hidden"
+              >
+                <time :datetime="time">
+                  {{ time }}
+                </time>
+                <span>{{ ` - ` }}</span>
+                <time :datetime="agendas[0].endTime">
+                  {{ agendas[0].endTime }}
+                </time>
+              </div>
+
+              <AgendaCard
+                v-for="(agenda, index) in agendas"
+                :key="`${time}-${index}`"
+                :data="agenda"
+                :index="index"
+                :time="time"
+                :show-time="index === 0"
+                :is-selected="
+                  agenda.tags?.some((tag) => selectedTags.includes(tag))
+                    || selectedTags.length === 0
+                "
+              />
+            </div>
+          </transition-group>
+
+          <transition-group
+            name="agenda-fade"
+            tag="div"
+            class="col-span-3"
+            appear
+          >
+            <div
+              v-if="selectedDate === '13'"
+              class="relative"
+            >
+              <time
+                class="sticky top-[99px] z-[5] flex h-7 w-full items-center bg-webconf-gray px-5 text-btn-16 text-webconf-blue lg:top-[124px] lg:ml-[-86px] lg:w-[86px] lg:justify-center"
+              >
+                <span>17:00</span>
+                <span class="lg:hidden">{{ ` - ` }}</span>
+                <span class="lg:hidden">17:10</span>
+              </time>
+
+              <h2
+                key="opening-speech"
+                class="col-span-3 border-b-[0.5px] border-webconf-gray/50 bg-black px-5 py-7 text-h4-24 lg:py-10 lg:text-center xl:py-14"
+              >
+                閉幕致詞
+              </h2>
+            </div>
+          </transition-group>
+        </div>
       </section>
     </main>
 
-    <!-- 提示 -->
-    <section class="py-[120px] text-center text-webconf-gray">
-      <p class="text-h5-20">
-        \ 更多精彩議程即將釋出 /
-      </p>
-    </section>
-
     <!-- 講者資訊彈跳視窗 -->
-    <!-- <AgentdaSpeakersPopver
-      v-if="isShowPopover"
-      @close="isShowPopover = false"
-    /> -->
+    <NuxtPage />
   </div>
 </template>
 
@@ -454,9 +424,30 @@ watch(
   background-position: center top;
   background-size: cover;
 }
+
 @media (min-width: 640px) {
   .agenda-section {
     background-position: center;
   }
+}
+
+/* 議程淡入淡出效果 */
+.agenda-fade-enter-active,
+.agenda-fade-leave-active {
+  transition: all 0.3s ease-in-out;
+}
+
+.agenda-fade-enter-from,
+.agenda-fade-leave-to {
+  opacity: 0;
+}
+
+.agenda-fade-enter-to,
+.agenda-fade-leave-from {
+  opacity: 1;
+}
+
+.agenda-fade-move {
+  transition: opacity 0.3s ease-in-out;
 }
 </style>
