+++ pages/sponsors.vue
