+++ public/_robots.txt
@@ -1,6 +1,1 @@
 User-Agent: *
-Disallow: /history/v2
-Disallow: /speakers/v2
-Disallow: /agenda/v2
-Disallow: /sponsors/v2
-Disallow: /staff/v2
